login and registartion changes

+++ posts/views.py
@@ -19,13 +19,16 @@ from .forms import PostForm
 
 
 def posts_create(request):
-	if not request.user.is_staff or not request.user.is_superuser:
+	# if not request.user.is_staff or not request.user.is_superuser:
+	if not request.user.is_authenticated:
 		raise Http404
 
 	form = PostForm(request.POST or None, request.FILES or None)
 
 	if form.is_valid():
+
 		instance = form.save(commit=False)
+		instance.user = request.user
 		print(form.cleaned_data.get('title'))
 		instance.save()
 		#message success
@@ -77,11 +80,15 @@ def posts_list(request):
 	return render(request, "post_list.html", context)
 
 def posts_update(request, slug=None):
-	# if not request.user.is_staff or not request.user.is_superuser:
-	# 	raise Http404
-
-	# if not request.user.is_authenticated():
-	# 	raise Http404
+	instance = get_object_or_404(Post, slug=slug)
+	if not request.user.is_staff or not request.user.is_superuser:
+		messages.success(request, "You don't have permission to do that.")
+		return HttpResponseRedirect(instance.get_absolute_url())
+	print("OK")
+	if instance.user != request.user:
+		print("hmm")
+		messages.success(request, "You don't have permission to edit someone else's post.")
+		return HttpResponseRedirect(instance.get_absolute_url())
 
 	instance = get_object_or_404(Post, slug=slug)
 	form = PostForm(request.POST or None, request.FILES or None, instance=instance)
@@ -101,6 +108,12 @@ def posts_update(request, slug=None):
 
 def post_delete(request, slug=None):
 	instance = get_object_or_404(Post, slug=slug)
+
+	if instance.user != request.user:
+		print("hmm")
+		messages.success(request, "You don't have permission to delete someone else's post.")
+		return HttpResponseRedirect(instance.get_absolute_url())
+
 	instance.delete()
 	messages.success(request, "Successfully Deleted")
 	return redirect('posts:list')
@@ -193,6 +206,7 @@ def posts_detail(request, slug):
 									parent=parent_obj
 									
 							)
+		messages.success(request, 'Your comment was added!')
 		return HttpResponseRedirect(new_comment.content_object.get_absolute_url())
 		
 	comments = instance.comments
